Move Badger use into internal cache package
